refactor(architecture): introduce AbstractClient and HttpTransport (#13, #14)
AbstractClient (CNIC namespace) is the shared foundation for all
registrar API clients. It holds every property and method previously
in CNR\Client that is not CNR-specific: configuration loading,
transport setup, debug mode, user agent, proxy/referer, credentials,
IDN conversion, OTE/LIVE switching, settings/isOTE accessors, and the
executeCurl() bridge. Three abstract methods define the contract each
concrete client must satisfy: request(), newSocketConfig(), and
setDefaultLogger().

HttpTransport (CNIC namespace) owns the cURL handle lifecycle and
exposes a single post() method. The $chandle property is removed from
the client layer entirely; close() now delegates to the transport.

CNR\Client extends AbstractClient and retains only what is CNR-specific:
request() with CNR command formatting, requestNextResponsePage(), and
requestAllResponsePages().

IBS\Client now extends AbstractClient directly instead of CNR\Client,
correctly reflecting that IBS shares the common foundation but not the
CNR-specific layer. The redundant constructor override is eliminated
since AbstractClient already calls newSocketConfig() during construction.
