+++ src/CNR/Client.php
@@ -9,12 +9,11 @@ declare(strict_types=1);
 
 namespace CNIC\CNR;
 
+use CNIC\AbstractClient;
 use CNIC\CNR\Logger as L;
 use CNIC\CNR\Response;
 use CNIC\CNR\SocketConfig;
 use CNIC\CommandFormatter;
-use CNIC\IDNA\Factory\ConverterFactory;
-use CNIC\LoggerInterface;
 
 /**
  * CNR API Client
@@ -22,366 +21,33 @@ use CNIC\LoggerInterface;
  * @psalm-api
  * @package CNIC\CNR
  */
-class Client
+class Client extends AbstractClient
 {
     /**
-     * context data for the client
-     * @var array<string,mixed>
+     * Instantiate CNR SocketConfig
      */
-    protected array $context = [];
-    /**
-     * registrar api settings
-     * @var array<mixed>
-     */
-    protected array $settings;
-
-    /**
-     * API connection url
-     */
-    protected string $socketURL;
-
-    /**
-     * Object covering API connection data
-     */
-    protected SocketConfig $socketConfig;
-
-    /**
-     * activity flag for debug mode
-     */
-    protected bool $debugMode;
-
-    /**
-     * user agent
-     */
-    protected string $ua = '';
-
-    /**
-     * additional curl options to use
-     * @var array<string>
-     */
-    protected array $curlopts = [];
-
-    /**
-     * logger function name for debug mode
-     */
-    protected LoggerInterface $logger;
-
-    /**
-     * is connected to OT&E
-     */
-    protected bool $isOTE = false;
-
-    /**
-     * curl handle cache
-     */
-    protected ?\CurlHandle $chandle = null;
-
-    /**
-     * Constructor
-     *
-     * @param string $path Path to the configuration file
-     */
-    public function __construct(string $path = "")
+    #[\Override]
+    protected function newSocketConfig(): SocketConfig
     {
-        $contents = file_get_contents($path);
-        if ($contents === false) {
-            $contents = "";
-        }
-        $settings = json_decode($contents, true);
-        if (is_null($settings) || $settings === false || $settings === true) {
-            $settings = [];
-        }
-        $this->settings = $settings;
-        $this->socketURL = "";
-        $this->debugMode = false;
-        $this->ua = "";
-        $this->socketConfig = new SocketConfig($this->settings["parameters"] ?? []);
-        $this->useLIVESystem();
-        $this->setDefaultLogger();
+        return new SocketConfig($this->settings["parameters"] ?? []);
     }
 
     /**
-     * set custom logger to use instead of default one
-     * create your own class implementing \CNIC\LoggerInterface
+     * Set default CNR logger
      * @return $this
      */
-    public function setCustomLogger(LoggerInterface $customLogger)
-    {
-        $this->logger = $customLogger;
-        return $this;
-    }
-
-    /**
-     * set default logger to use
-     * @return $this
-     */
-    public function setDefaultLogger()
+    #[\Override]
+    public function setDefaultLogger(): static
     {
         $this->logger = new L();
         return $this;
     }
 
     /**
-     * Enable Debug Output to STDOUT
-     * @return $this
-     */
-    public function enableDebugMode()
-    {
-        $this->debugMode = true;
-        return $this;
-    }
-
-    /**
-     * Disable Debug Output
-     * @return $this
-     */
-    public function disableDebugMode()
-    {
-        $this->debugMode = false;
-        return $this;
-    }
-
-    /**
-     * Serialize given command for POST request including connection configuration data
-     * @param array<string,mixed> $cmd API command to encode
-     * @param bool $secured secure password (when used for output)
-     */
-    public function getPOSTData(array $cmd, bool $secured = false): string
-    {
-        return $this->socketConfig->getPOSTData($cmd, $secured);
-    }
-
-    /**
-     * Get the API Session ID that is currently set
-     */
-    public function getSession(): ?string
-    {
-        $sessid = $this->socketConfig->getSession();
-        return ($sessid === "" ? null : $sessid);
-    }
-
-    /**
-     * Get the API connection url that is currently set
-     */
-    public function getURL(): string
-    {
-        return $this->socketURL;
-    }
-
-    /**
-     * Set a custom user agent (for platforms that use this SDK)
-     * @param string $str user agent label
-     * @param string $rv user agent revision
-     * @param array<string> $modules further modules to add to user agent string, format: ["<module1>/<version>", "<module2>/<version>", ... ]
-     * @return $this
-     */
-    public function setUserAgent(string $str, string $rv, array $modules = [])
-    {
-        $mods = $modules === [] ? "" : " " . implode(" ", $modules);
-        $this->ua = $str . " (" . PHP_OS . "; " . php_uname("m") . "; rv:" . $rv . ")" . $mods . " php-sdk/" . $this->getVersion() . " php/" . implode(".", [PHP_MAJOR_VERSION, PHP_MINOR_VERSION, PHP_RELEASE_VERSION]);
-        return $this;
-    }
-
-    /**
-     * Get the user agent string
-     */
-    public function getUserAgent(): string
-    {
-        if ($this->ua === '') {
-            $this->ua = "PHP-SDK (" . PHP_OS . "; " . php_uname("m") . "; rv:" . $this->getVersion() . ") php/" . implode(".", [PHP_MAJOR_VERSION, PHP_MINOR_VERSION, PHP_RELEASE_VERSION]);
-        }
-        return $this->ua;
-    }
-
-    /**
-     * Set proxy to use for API communication
-     * @param string $proxy proxy to use (optional, for reset)
-     * @return $this
-     */
-    public function setProxy(string $proxy = "")
-    {
-        if ($proxy === '' || $proxy === '0') {
-            unset($this->curlopts[CURLOPT_PROXY]);
-        } else {
-            $this->curlopts[CURLOPT_PROXY] = $proxy;
-        }
-        return $this;
-    }
-
-    /**
-     * Get proxy configuration for API communication
-     */
-    public function getProxy(): ?string
-    {
-        if (isset($this->curlopts[CURLOPT_PROXY])) {
-            return $this->curlopts[CURLOPT_PROXY];
-        }
-        return null;
-    }
-
-    /**
-     * Set Referer to use for API communication
-     * @param string $referer Referer (optional, for reset)
-     * @return $this
-     */
-    public function setReferer(string $referer = "")
-    {
-        if ($referer === '' || $referer === '0') {
-            unset($this->curlopts[CURLOPT_REFERER]);
-        } else {
-            $this->curlopts[CURLOPT_REFERER] = $referer;
-        }
-        return $this;
-    }
-
-    /**
-     * Get Referer configuration for API communication
-     */
-    public function getReferer(): ?string
-    {
-        if (isset($this->curlopts[CURLOPT_REFERER])) {
-            return $this->curlopts[CURLOPT_REFERER];
-        }
-        return null;
-    }
-
-    /**
-     * Get the current module version
-     */
-    public function getVersion(): string
-    {
-        return "14.1.2";
-    }
-
-    /**
-     * Set another connection url to be used for API communication
-     * @param string $value API connection url to set
-     * @return $this
-     */
-    public function setURL(string $value)
-    {
-        $this->socketURL = $value;
-        return $this;
-    }
-
-    /**
-     * Set an API session id to be used for API communication
-     * @param string $value API session id (optional, for reset)
-     * @return $this
-     */
-    public function setSession(string $value = "")
-    {
-        $this->socketConfig->setSession($value);
-        return $this;
-    }
-
-    /**
-     * Set an Remote IP Address to be used for API communication
-     * To be used in case you have an active ip filter setting.
-     * @param string $value Remote IP Address (optional, for reset)
-     * @throws \Exception in case this feature is unsupported
-     * @return $this
-     */
-    public function setRemoteIPAddress(string $value = "")
-    {
-        if ($value !== '' && $value !== '0' && !isset($this->settings["parameters"]["ipfilter"])) {
-            throw new \Exception("Feature `IP Filter` not supported");
-        }
-        $this->socketConfig->setRemoteAddress($value);
-        return $this;
-    }
-
-    /**
-     * Set Credentials to be used for API communication
-     * @param string $uid account name (optional, for reset)
-     * @param string $pw account password (optional, for reset)
-     * @return $this
-     */
-    public function setCredentials(string $uid = "", string $pw = "")
-    {
-        $this->socketConfig->setLogin($uid);
-        $this->socketConfig->setPassword($pw);
-        return $this;
-    }
-
-    /**
-     * Set Credentials to be used for API communication
-     * @param string $uid account name (optional, for reset)
-     * @param string $role role user id (optional, for reset)
-     * @param string $pw role user password (optional, for reset)
-     * @return $this
-     */
-    public function setRoleCredentials(string $uid = "", string $role = "", string $pw = "")
-    {
-        $login = $uid;
-        if ($role !== '' && $role !== '0') {
-            $login .= $this->settings["roleSeparator"] . $role;
-        }
-        return $this->setCredentials($login, $pw);
-    }
-
-    /**
-     * Convert domain names to idn + punycode if necessary
-     * @param array<string> $domains list of domain names (or tlds)
-     * @return array<mixed>
-     */
-    public function IDNConvert(array $domains): array
-    {
-        return ConverterFactory::convert($domains);
-    }
-
-    /**
-     * Auto convert API command parameters to punycode, if necessary.
-     *
-     * @param array<string> $cmd API command
-     * @return array<string>
-     */
-    protected function autoIDNConvert(array $cmd): array
-    {
-        // only convert if configured for the registrar
-        // and ignore commands in string format (even deprecated)
-        if (
-            !$this->settings["needsIDNConvert"]
-            || !function_exists("idn_to_ascii")
-        ) {
-            return $cmd;
-        }
-
-        $asciipattern = "/^[a-zA-Z0-9\.-]+$/i";
-        // DOMAIN params get auto-converted by API
-        // RSRBE-7149 for NS coverage
-        $keypattern = "/^(PARENTDOMAIN|NAMESERVER|NS|DNSZONE)([0-9]*)$/i";
-        $objclasspattern = "/^(DOMAIN(APPLICATION|BLOCKING)?|NAMESERVER|NS|DNSZONE)$/i";
-        $toconvert = [];
-        $idxs = [];
-        foreach ($cmd as $key => $val) {
-            if (
-                ((bool)preg_match($keypattern, $key)
-                    // RSRTPM-3167: OBJECTID is a PATTERN in CNR API and not supporting IDNs
-                    || ($key === "OBJECTID"
-                        && isset($cmd["OBJECTCLASS"])
-                        && (bool)preg_match($objclasspattern, $cmd["OBJECTCLASS"])
-                    )
-                )
-                && !(bool)preg_match($asciipattern, $val)
-            ) {
-                $toconvert[] = $val;
-                $idxs[] = $key;
-            }
-        }
-        if ($toconvert !== []) {
-            $results = $this->IDNConvert($toconvert);
-            foreach ($results as $idx => $row) {
-                $cmd[$idxs[$idx]] = $row["punycode"];
-            }
-        }
-        return $cmd;
-    }
-
-    /**
      * Perform API request using the given command
      * @param array<mixed> $cmd API command to request (optional for session login)
      */
+    #[\Override]
     public function request(array $cmd = []): Response
     {
         $mycmd = CommandFormatter::flattenCommand($cmd);
@@ -397,52 +63,7 @@ class Client
     }
 
     /**
-     * Initialise the cURL handle if needed, configure options, execute and return the result.
-     * @param string $data serialized POST payload
-     * @param array<string> $cfg connection config (must contain CONNECTION_URL)
-     * @param array<int, mixed> $extraCurlOpts additional cURL options merged over the defaults
-     * @return array{0: string, 1: string|null} [rawResponse, errorMessage|null]
-     */
-    protected function executeCurl(string $data, array $cfg, array $extraCurlOpts = []): array
-    {
-        if (!$this->chandle instanceof \CurlHandle) {
-            $tmp = curl_init();
-            if ($tmp === false) {
-                return ["nocurl", "CURL for PHP missing."];
-            }
-            $this->chandle = $tmp;
-        }
-
-        curl_setopt_array($this->chandle, [
-            // CURLOPT_VERBOSE         => $this->debugMode,
-            CURLOPT_URL             => $cfg["CONNECTION_URL"],
-            CURLOPT_CONNECTTIMEOUT  => 30, // 30s, 300s by default
-            CURLOPT_TIMEOUT         => $this->settings["socketTimeout"],
-            CURLOPT_POST            => 1,
-            CURLOPT_HEADER          => 0,
-            CURLOPT_RETURNTRANSFER  => 1,
-            CURLOPT_POSTFIELDS      => $data,
-            CURLOPT_USERAGENT       => $this->getUserAgent(),
-            CURLOPT_HTTPHEADER      => [
-                "Expect:",
-                "Content-Type: application/x-www-form-urlencoded", //UTF-8 implied
-                "Content-Length: " . (string)strlen($data),
-                "Connection: keep-alive"
-            ]
-        ] + $extraCurlOpts + $this->curlopts);
-
-        $r = curl_exec($this->chandle);
-        \assert(\is_string($r) || $r === false);
-        if ($r === false) {
-            $error = curl_error($this->chandle);
-            return ["httperror|" . $error, $error];
-        }
-        return [$r, null];
-    }
-
-    /**
      * Request the next page of list entries for the current list query
-     * Useful for tables
      * @param Response $rr API Response of current page
      * @throws \Exception in case Command Parameter LAST is in use while using this method
      */
@@ -464,7 +85,6 @@ class Client
             $mycmd["LIMIT"] = $limit;
             return $this->request($mycmd);
         }
-
         return null;
     }
 
@@ -484,91 +104,5 @@ class Client
             $tmp = $this->requestNextResponsePage($tmp);
         } while ($tmp instanceof Response);
         return $responses;
-    }
-
-    /**
-     * Close all curl handles
-     */
-    public function close(): void
-    {
-        if (!is_null($this->chandle)) {
-            curl_close($this->chandle);
-        }
-    }
-
-    /**
-     * Set a data view to a given subuser
-     * @param string $uid subuser account name
-     * @return $this
-     */
-    public function setUserView(string $uid = "")
-    {
-        $this->socketConfig->setUser($uid);
-        return $this;
-    }
-
-    /**
-     * Activate High Performance Setup
-     * @return $this
-     */
-    public function useHighPerformanceConnectionSetup()
-    {
-        $oldurl = $this->getURL();
-        $hostname = parse_url($oldurl, PHP_URL_HOST);
-        if (is_string($hostname) && $hostname !== '') {
-            $url = str_replace($hostname, "127.0.0.1", $oldurl);
-            $url = str_replace("https://", "http://", $url);
-            $this->setURL($url);
-        }
-        return $this;
-    }
-
-    /**
-     * Get registrar API settings
-     * @return array<mixed>
-     */
-    public function getSettings(): array
-    {
-        return $this->settings;
-    }
-
-    /**
-     * Check whether the client is connected to the OT&E system
-     */
-    public function isOTE(): bool
-    {
-        return $this->isOTE;
-    }
-
-    /**
-     * Set OT&E System for API communication
-     * @return $this
-     */
-    public function useOTESystem()
-    {
-        $this->isOTE = true;
-        return $this->setURL($this->settings["env"]["ote"]["url"]);
-    }
-
-    /**
-     * Set LIVE System for API communication (this is the default setting)
-     * @return $this
-     */
-    public function useLIVESystem()
-    {
-        $this->isOTE = false;
-        return $this->setURL($this->settings["env"]["live"]["url"]);
-    }
-    /**
-     * Set context data for the client (to be used in a custom logger to access additional data at logging time again)
-     * The use of this context data is optional and has no impact on the SDK behaviour itself.
-     * Will be forwarded either to the logger or to the Response object
-     * @param array<string,mixed> $context
-     * @return $this
-     */
-    public function setContext(array $context)
-    {
-        $this->context = $context;
-        return $this;
     }
 }
